rename model names (#7)
