+++ models/updateComment_model.js
@@ -1,6 +1,6 @@
 const mongoose = require("mongoose");
 
-const CourseRate = require("./rating_model");
+const CourseRate = require("./schema/rating_model");
 
 module.exports = function uploadComments(findObj,dataObj) {
   //console.log("=",req.params)
